[Feature][cdc] UT
